% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/about_me.R
\name{about_me}
\alias{about_me}
\title{Title A little about me}
\usage{
about_me(my_name, birth_year, fav_animal)
}
\arguments{
\item{my_name}{a character containing your name}

\item{birth_year}{A number containing your birth year}

\item{fav_animal}{a character containing your favorite animal species}
}
\value{

}
\description{
some cools description detail about me including my name, birth year, and favorite animal.
}
