% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakfast_laundry.R
\name{breakfast_time}
\alias{breakfast_time}
\title{Breakfast plan}
\usage{
breakfast_time(food, num, day)
}
\arguments{
\item{food}{a character value describing favorite breakfast food}

\item{num}{a numeric value describing quanitity of items consumed}

\item{day}{a character describing day of the week}
}
\value{

}
\description{
Describes the about of food one should eat for breakfast each day
}
