% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakfast_laundry.R
\name{laundry_elec_cost}
\alias{laundry_elec_cost}
\title{Cost of doing laundry function}
\usage{
laundry_elec_cost(kw_hr, frequency, cost, weeks)
}
\arguments{
\item{kw_hr}{a numeric value of the cost of electricity}

\item{frequency}{a numeric value on how many times per week you plan to do laundry}

\item{cost}{a numeric value on how much it costs to do a single load of lanudry (assuming coin-op)}

\item{weeks}{a numeric value describing the number of weeks you plan to do lanudry}
}
\value{

}
\description{
Provides an overview of how much money it costs to run a laundry machine.  Inputs include electrity cost (kw/h), frequency of laundry per week, cost per each load, and number of weeks you plan to do laundry
}
