% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space_man.R
\name{space_man}
\alias{space_man}
\title{Title Spaceman comes to destry earth}
\usage{
space_man(planet, future_year, fav_burrito)
}
\arguments{
\item{planet}{insert a character name on some planet}

\item{future_year}{insert a numeric data of some time in the future}

\item{fav_burrito}{insert a character describing your favorite burrito}
}
\value{

}
\description{
Title Spaceman comes to destry earth
}
